--- src/app/components/cats/state/cats.state.ts
@@ -1,11 +1,0 @@
-import { Cat } from 'src/app/interfaces/Cat';
-
-export interface CatsState {
-  cats: Cat[];
-  loading: boolean;
-}
-
-export const initialState: CatsState = {
-  cats: [],
-  loading: true,
-};
