--- src/app/components/cats/state/index.ts
@@ -1,4 +1,0 @@
-export { reducer } from './cats.reducer';
-export * from './cats.actions';
-export { CatsState, initialState } from './cats.state';
-export { CatsEffects } from './cats.effects';
