+++ src/app/store/index.ts
@@ -1,0 +1,3 @@
+export { reducer, CatsState, initialState } from './reducers/cats.reducer';
+export * from './actions/cats.actions';
+export { CatsEffects } from './effects/cats.effects';
