+++ src/app/store/reducers/cats.reducer.ts
@@ -1,6 +1,16 @@
 import { Action, createReducer, on } from '@ngrx/store';
-import * as catsActions from './cats.actions';
-import { CatsState, initialState } from './cats.state';
+import * as catsActions from '../actions/cats.actions';
+import { Cat } from 'src/app/interfaces/Cat';
+
+export interface CatsState {
+  cats: Cat[];
+  loading: boolean;
+}
+
+export const initialState: CatsState = {
+  cats: [],
+  loading: true,
+};
 
 const catsReducer = createReducer(
   initialState,
